LRCI-3821 Move status checking logic into DefaultBuildRun to leave room for Jethr0 status checking
